bugfixed WHERE_LT added jmp tests added two mutexes: one for instruction+reduction and one for ios. Why do we need same mutex for instruction and reduction ?
git-svn-id: https://apollo.arh.pub.ro:8443/svn/ConnexSIFT/trunk/software/connex@372 0e9f8c0c-05a5-1044-940b-b810a0af9d24

+++ simulator/src/ConnexVector.cpp
@@ -39,7 +39,7 @@ ConnexVector ConnexVector::operator op(ConnexVector anotherVector)	\
         else eqFlag.cells[i] = 0;
 
 #define BINARY_OP_FLAGS_LT(op)                                      \
-		if (cells[i] < anotherVector.cells[i])                      \
+		if ((short)cells[i] < (short)anotherVector.cells[i])                      \
             ltFlag.cells[i] = 1;                                    \
         else ltFlag.cells[i] = 0;
